% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_PK_model_fxns.R
\name{GR_inVitro_integration}
\alias{GR_inVitro_integration}
\title{GR_inVitro_integration
Calculate relative tumor growth inhibition (TGR) based on conc_profile}
\usage{
GR_inVitro_integration(conc_profile, GR_para, int_method = "mean")
}
\arguments{
\item{conc_profile}{data.frame with \code{Time} (in days) and relative growth rate \code{k} over time
can be the output of \code{PK_to_conc_profile}}

\item{GR_para}{numeric array for the in vitro GR parameters of the drug (required fields: GR_inf, GEC50, h_GR)
Can calculate the average of multiple parameters at the same time given multiple rows in \code{GR_para}}

\item{int_method}{string to define the type of integration}
}
\value{
numeric value for relative tumor growth rate (TGR)
}
\description{
Calculate the serum concentration profile based on the drug PK, dose, and schedule.
The concentration (in µM) is calculated based on one-compartment PK model with parameters k_a and k_e
provided in the input variable PK_para.
}
\examples{

## regular QD treatment

PK_para = data.frame(k_a = 2, k_e = 0.25, MW = 450, VF = 1)
GR_para = data.frame(GR_inf = -0.5, GEC50 = 0.3, h_GR = 1.5)
Schedule = "QD"
Duration = 21
Dose = 5
conc_profile = PK_to_conc_profile(PK_para, Dose, Schedule, Duration)
relk_1 = relk_over_time(GR_para, conc_profile)
GR_predicted = GR_inVitro_integration(conc_profile, GR_para)
}
