% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{logisticFit}
\alias{logisticFit}
\title{Actual fitting function}
\usage{
logisticFit(
  concs,
  normValues,
  x_0 = 1,
  curve_type = c("IC", "GR"),
  force = FALSE,
  cap = 0.1
)
}
\arguments{
\item{concs}{concentration values}

\item{normValues}{normalized response values}

\item{x_0}{upper limit; (=1 by default)}

\item{curve_type}{response curve: either IC (for values in \code{[0,1]}) or GR (for values in \code{[-1,1]})}

\item{force}{force a sigmoidal fit even if the fit is not significantly better than a flat fit}

\item{cap}{cap values at (x_0 + cap)}
}
\value{
vector of parameters
}
\description{
\code{logisticFit} returns fit parameters for an IC or GR curve
}
\details{
returns fit parameters
}
