% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_PK_model_fxns.R
\name{PK_to_conc_profile}
\alias{PK_to_conc_profile}
\title{PK_to_conc_profile
Calculate the serum concentration profile based on the drug PK, dose, and schedule}
\usage{
PK_to_conc_profile(PK_para, Dose, Schedule, Duration, Dose_uM_0 = 0)
}
\arguments{
\item{PK_para}{numeric array for the PK properties of the drug (required fields: MW, VF, k_a, k_e)
\itemize{
  \item{'MW' molecular weight in g/mol}
  \item{'VF' distribution volume in L/kg}
  \item{'k_a' absoption rate in 1/h}
  \item{'k_e' elimination rate in 1/h}
  }}

\item{Dose}{numeric values for the dose of the drug given in mg/kg}

\item{Schedule}{string for the frequency of the treatment given (QC, BID or EOD)}

\item{Duration}{numeric value for the duration of the treatment (in days)}

\item{Dose_uM_0}{numeric value for the initial serum concentration (in µM)
Defaults to \code{0}.}
}
\value{
data.frame with \code{Time} (in days) and \code{Conc} serum concentration (in µM)
}
\description{
Calculate the serum concentration profile based on the drug PK, dose, and schedule.
The concentration (in µM) is calculated based on one-compartment PK model with parameters k_a and k_e 
provided in the input variable PK_para.
}
