% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_PK_model_fxns.R
\name{relk_over_time}
\alias{relk_over_time}
\title{relk_over_time
Calculate the serum concentration profile based on the drug PK, dose, and schedule}
\usage{
relk_over_time(GR_para, conc_profile)
}
\arguments{
\item{GR_para}{numeric array for the in vitro GR parameters of the drug (required fields: \code{GR_inf}, \code{GEC50}, \code{h_GR})}

\item{conc_profile}{data.frame with \code{Time} (in days) and \code{Conc} (serum concentration in µM)
can be the output of PK_to_conc_profile}
}
\value{
data.frame with\code{Time} (in days) and relative growth rate \code{k} over time
}
\description{
Calculate the serum concentration profile based on the drug PK, dose, and schedule.
The concentration (in µM) is calculated based on one-compartment PK model with parameters k_a and k_e
provided in the input variable PK_para.
}
