% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGRmodel-package.R
\docType{package}
\name{TGRmodel-package}
\alias{TGRmodel}
\alias{TGRmodel-package}
\title{TGRmodel: Functions to predict Tumor Growth Rate inhibition using in vitro data}
\description{
Functions to predict Tumor Growth Rate inhibition using in vitro data.
    Model described in the manuscript "Growth-rate model predicts in vivo tumor response from in vitro data" (2022, CPT: Pharmacometrics & Systems Pharmacology).
    Please cite "Diegmiller et al., in revision at CPT Pharmacometrics Syst Pharmacol, 2022".
}
\keyword{internal}
